* More warnings killed
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8703 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/codecs/cmml/dsfCMMLRawSource/CMMLRawSourcePin.cpp
@@ -173,7 +173,7 @@ HRESULT CMMLRawSourcePin::deliverTag(C_CMMLTag* inTag) {
 
 	debugLog<<"Tag time = "<<locStart<<" - "<<locStop<<endl;
 	debugLog<<locNarrowStr<<endl<<endl;
-	locSample->SetActualDataLength(locNarrowStr.size());
+	locSample->SetActualDataLength((long)locNarrowStr.size());
 	locSample->SetTime(&locStart, &locStop);
 	locSample->SetMediaTime(NULL, NULL);
 	locSample->SetDiscontinuity(FALSE);
